Introduce foreign table service
