+++ packages/twenty-server/src/engine/metadata-modules/remote-server/remote-table/foreign-table/utils/get-foreign-table-column-name.util.ts
